requirements.txt
absl-py==2.1.0
affine==2.4.0
aiofiles==23.2.1
albucore==0.0.12
albumentations==0.5.2
altair==5.3.0
annotated-types==0.6.0
anyio==4.3.0
argcomplete==3.3.0
asttokens==2.4.1
attrs==23.2.0
backcall==0.2.0
beautifulsoup4==4.11.1
blinker==1.7.0
Bottleneck @ file:///C:/Windows/Temp/abs_3198ca53-903d-42fd-87b4-03e6d03a8381yfwsuve8/croots/recipe/bottleneck_1657175565403/work
brotlipy @ file:///D:/bld/brotlipy_1648854327487/work
cached-property @ file:///home/conda/feedstock_root/build_artifacts/cached_property_1615209429212/work
cachetools==5.3.3
cadquery @ git+https://github.com/CadQuery/cadquery.git@e8d74474ca4869d014cba6e1f835cdabad86bf88
cadquery-ocp==7.7.1
casadi==3.6.3
certifi @ file:///C:/b/abs_91u83siphd/croot/certifi_1700501720658/work/certifi
cffi @ file:///D:/bld/cffi_1656782930891/work
charset-normalizer @ file:///home/conda/feedstock_root/build_artifacts/charset-normalizer_1661170624537/work
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
cloudpickle @ file:///C:/b/abs_3796yxesic/croot/cloudpickle_1683040098851/work
colorama==0.4.5
coloredlogs==15.0.1
colorlog==6.8.2
colormath==3.0.0
comm==0.2.2
cos-python-sdk-v5==1.9.31
crcmod==1.7
cryptography @ file:///C:/ci/cryptography_1652083563162/work
cssselect==1.2.0
cycler @ file:///home/conda/feedstock_root/build_artifacts/cycler_1635519461629/work
Cython==3.0.3
debugpy==1.8.1
decorator==5.1.1
distinctipy==1.3.4
distlib==0.3.8
easydict==1.10
einops==0.4.1
et-xmlfile==1.1.0
eval_type_backport==0.2.0
exceptiongroup==1.2.0
executing==2.0.1
ezdxf==1.0.3
fastapi==0.110.1
ffmpy==0.3.2
filelock==3.13.4
Flask==3.0.3
flatbuffers==24.3.25
fonttools @ file:///D:/bld/fonttools_1661381554081/work
fsspec==2024.3.1
future==0.18.3
gdown==4.5.3
gitdb==4.0.11
GitPython==3.1.43
gradio==3.41.0
gradio_client==0.5.0
grpcio==1.65.4
GuideConv==0.0.0
h11==0.14.0
h5py @ file:///C:/b/abs_17fav01gwy/croot/h5py_1691589733413/work
hdf5storage==0.1.19
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.22.2
humanfriendly==10.0
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1642433548627/work
imageio==2.34.2
imgaug==0.4.0
imgvision==0.1.7.1
importlib_metadata==7.1.0
importlib_resources==6.4.0
iniconfig==2.0.0
ipykernel==6.29.4
ipython==8.18.1
iso8601==1.1.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.3
joblib @ file:///C:/b/abs_1anqjntpan/croot/joblib_1685113317150/work
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.1
jupyter_core==5.7.2
kaleido==0.2.1
kiwisolver @ file:///D:/bld/kiwisolver_1657953200729/work
lazy_loader==0.4
llvmlite @ file:///C:/b/abs_458aeunq7a/croot/llvmlite_1697031108211/work
lxml==5.2.2
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib @ file:///D:/bld/matplotlib-suite_1661439961257/work
matplotlib-inline==0.1.7
mdurl==0.1.2
mkl-fft @ file:///C:/b/abs_19i1y8ykas/croot/mkl_fft_1695058226480/work
mkl-random @ file:///C:/b/abs_edwkj1_o69/croot/mkl_random_1695059866750/work
mkl-service==2.4.0
mpmath==1.3.0
multimethod==1.9.1
munkres==1.1.4
nest-asyncio==1.6.0
networkx==2.8.6
nibabel==5.2.1
nlopt==2.7.1
nox==2024.4.15
nptyping==2.0.1
numba @ file:///C:/b/abs_2eee8kaps6/croot/numba_1701362407655/work
numexpr @ file:///C:/b/abs_5fucrty5dc/croot/numexpr_1696515448831/work
numpy @ file:///C:/b/abs_16b2j7ad8n/croot/numpy_and_numpy_base_1704311752418/work/dist/numpy-1.26.3-cp39-cp39-win_amd64.whl#sha256=02e606e23ca31bb00a40d147fd1ce4dd7d241395346a4196592d5abe54a333bc
onnxruntime==1.18.1
opencv-python==4.6.0.66
opencv-python-headless==4.10.0.84
openpyxl==3.1.2
orjson==3.10.0
outcome==1.3.0.post0
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1637239678211/work
pandas==1.5.0
parso==0.8.4
path==16.7.1
patsy==0.5.3
peewee==3.17.6
Pillow==9.4.0
platformdirs==4.2.0
plotly==5.10.0
plotly-express==0.4.1
pluggy==1.4.0
ply==3.11
pooch==1.8.1
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.8
ptwt==0.1.8
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==15.0.2
pycocotools==2.0.8
pycparser @ file:///home/conda/feedstock_root/build_artifacts/pycparser_1636257122734/work
pycryptodome==3.20.0
pydantic==2.7.0
pydantic_core==2.18.1
pydeck==0.8.1b0
pydub==0.25.1
Pygments==2.17.2
pynvml==11.5.3
PyOpenGL @ file:///C:/Users/15432/Desktop/PyOpenGL-3.1.5-cp39-cp39-win_amd64.whl
PyOpenGL-accelerate @ file:///C:/Users/15432/Desktop/PyOpenGL_accelerate-3.1.5-cp39-cp39-win_amd64.whl
pyOpenSSL @ file:///home/conda/feedstock_root/build_artifacts/pyopenssl_1643496850550/work
pyparsing @ file:///home/conda/feedstock_root/build_artifacts/pyparsing_1652235407899/work
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
pyquery==2.0.0
pyreadline3==3.4.1
pyrealsense2==2.54.1.5217
PySocks @ file:///D:/bld/pysocks_1661604991356/work
pytest==8.1.1
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1626286286081/work
python-multipart==0.0.9
pytz==2022.5
PyWavelets==1.3.0
pywin32==306
PyYAML==6.0
pyzed @ file:///C:/Users/15432/Desktop/pyzed-3.8-cp39-cp39-win_amd64.whl
pyzmq==26.0.2
rasterio==1.3.10
referencing==0.34.0
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1656534056640/work
rich==13.7.1
rpds-py==0.18.0
ruff==0.3.7
safetensors==0.4.3
scikit-image==0.24.0
scikit-learn @ file:///C:/b/abs_daon7wm2p4/croot/scikit-learn_1694788586973/work
scipy==1.11.4
seaborn @ file:///C:/b/abs_68ltdkoyoo/croot/seaborn_1673479199997/work
selenium==4.23.0
semantic-version==2.10.0
serial==0.0.97
shap @ file:///C:/b/abs_e214ia_u8b/croot/shap_1692866029183/work
shapely==2.0.5
shellingham==1.5.4
simplification==0.7.12
sip @ file:///D:/bld/sip_1655765083928/work
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
slicer @ file:///tmp/build/80754af9/slicer_1633422823758/work
smmap==5.0.1
sniffio==1.3.1
snuggs==1.4.7
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
spectral==0.22.4
stack-data==0.6.3
starlette==0.37.2
statsmodels==0.13.2
streamlit==1.33.0
swanboard==0.1.3b4
swankit==0.1.1b1
swanlab==0.3.16
sympy==1.13.1
tenacity==8.1.0
tensorboard==2.17.0
tensorboard-data-server==0.7.2
thop==0.1.1.post2209072238
threadpoolctl @ file:///Users/ktietz/demo/mc3/conda-bld/threadpoolctl_1629802263681/work
tifffile==2022.8.12
timm==1.0.7
toml @ file:///home/conda/feedstock_root/build_artifacts/toml_1604308577558/work
tomli==2.0.1
tomlkit==0.12.0
toolz==0.12.1
torch==1.12.1
torchaudio==0.12.1
torchvision==0.13.1
tornado @ file:///D:/bld/tornado_1656937934674/work
tqdm==4.64.1
traitlets==5.14.3
trio==0.26.0
trio-websocket==0.11.1
typer==0.12.3
typing_extensions==4.9.0
typish==1.9.3
ujson==5.10.0
ultralytics==8.2.75
ultralytics-thop==2.0.0
unicodedata2 @ file:///D:/bld/unicodedata2_1649112114189/work
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1658789158161/work
uvicorn==0.29.0
vega-datasets==0.9.0
virtualenv==20.25.3
watchdog==4.0.0
wcwidth==0.2.13
websocket-client==1.8.0
websockets==11.0.3
Werkzeug==3.0.3
win-inet-pton @ file:///D:/bld/win_inet_pton_1648771910527/work
wsproto==1.2.0
xmltodict==0.13.0
zipp==3.18.1
